// src/services/central-api/generated/model/ctl-mst-vehicle-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PL Cental API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CtlMstVehicleType
 */
export interface CtlMstVehicleType {
    /**
     * 
     * @type {string}
     * @memberof CtlMstVehicleType
     */
    'code'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstVehicleType
     */
    'type'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstVehicleType
     */
    'createTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstVehicleType
     */
    'createBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstVehicleType
     */
    'status'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstVehicleType
     */
    'updateTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstVehicleType
     */
    'updateBy'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CtlMstVehicleType
     */
    'isDelete'?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstVehicleType
     */
    'deleteBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstVehicleType
     */
    'deleteTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstVehicleType
     */
    'tenantCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstVehicleType
     */
    'branchCode'?: string | null;
}